+++ apps/landing/src/app/test-case/page.tsx
@@ -1,13 +1,12 @@
 import 'katex/dist/katex.min.css'
 
-import { Box, css, Flex, Grid, Text, VStack } from '@devup-ui/react'
+import { Box, css, Flex, Text, VStack } from '@devup-ui/react'
 import { readFile } from 'fs/promises'
 import { Metadata } from 'next'
-import Latex from 'react-syntax-highlighter/dist/cjs/languages/hljs/latex'
 
 import { FailedOnlyInput } from '@/components/test-case/FailedOnlyInput'
+import { TestCaseList } from '@/components/test-case/list/TestCaseList'
 import { TestCaseTable } from '@/components/test-case/table/TestCaseTable'
-import TestCaseCircle from '@/components/test-case/TestCaseCircle'
 import { TestCaseDisplayBoundary } from '@/components/test-case/TestCaseDisplayBoundary'
 import { TestCaseProvider } from '@/components/test-case/TestCaseProvider'
 import { TestCaseTypeBoundary } from '@/components/test-case/TestCaseTypeBoundary'
@@ -61,47 +60,7 @@ export default async function TestCasePage() {
             <TestCaseTable results={testStatus[key][2]} />
           </TestCaseTypeBoundary>
           <TestCaseTypeBoundary type="list">
-            <Grid
-              gap="8px"
-              gridTemplateColumns="repeat(auto-fill, minmax(16px, 1fr))"
-            >
-              {testStatus[key][2].map(
-                ([text, expected, actual, isSuccess], idx) => {
-                  const textParts = parseTextWithLaTeX(text)
-
-                  return (
-                    <TestCaseDisplayBoundary
-                      key={text + idx}
-                      display={!isSuccess}
-                      option="failedOnly"
-                    >
-                      <TestCaseCircle key={text + idx} isSuccess={isSuccess}>
-                        <Text
-                          color="#FFF"
-                          typography="body"
-                          whiteSpace="nowrap"
-                          wordBreak="keep-all"
-                        >
-                          {textParts.map((part, partIdx) =>
-                            part.type === 'latex' ? (
-                              <Latex key={partIdx}>${part.content}$</Latex>
-                            ) : (
-                              <span key={partIdx}>{part.content}</span>
-                            ),
-                          )}
-                          <br />
-                          정답 : {expected}
-                          <br />
-                          결과 : {actual}
-                          <br />
-                          {isSuccess ? '✅ 테스트 성공' : '❌ 테스트 실패'}
-                        </Text>
-                      </TestCaseCircle>
-                    </TestCaseDisplayBoundary>
-                  )
-                },
-              )}
-            </Grid>
+            <TestCaseList results={testStatus[key][2]} />
           </TestCaseTypeBoundary>
         </VStack>
       </TestCaseDisplayBoundary>
@@ -136,7 +95,7 @@ export default async function TestCasePage() {
             을 기반으로 작성되었습니다.
           </Text>
         </VStack>
-        <VStack gap="40px" px={['16px', null, null, '60px']}>
+        <VStack gap="12px" px={['16px', null, null, '60px']}>
           <Flex
             alignItems="center"
             color="$primary"
@@ -168,57 +127,10 @@ export default async function TestCasePage() {
               실패한 케이스만 표시하기
             </Text>
           </Flex>
-          <Box bg="$text" h="1px" />
         </VStack>
+        <Box bg="$text" h="1px" mt="40px" mx={['16px', null, null, '60px']} />
         {cases}
       </Box>
     </TestCaseProvider>
   )
-}
-
-/**
- * This function parses text with LaTeX expressions and returns an array of parts.
- * It assumes that LaTeX is wrapped in double dollar delimiters ($$...$$).
- * Note that single dollar delimiters ($...$) are not rendered.
- * @param input - The input text to parse.
- * @returns An array of parts, where each part is either a text or a LaTeX expression.
- */
-const parseTextWithLaTeX = (input: string) => {
-  const parts: Array<{
-    type: 'text' | 'latex'
-    content: string
-  }> = []
-  const latexRegex = /\$\$([^$]+(?:\$(?!\$)[^$]*)*)\$\$/g
-  let lastIndex = 0
-  let match
-
-  while ((match = latexRegex.exec(input)) !== null) {
-    // if there is text before the LaTeX expression, add it as a text part:
-    if (match.index > lastIndex) {
-      const textContent = input.slice(lastIndex, match.index)
-      if (textContent) {
-        parts.push({ type: 'text', content: textContent })
-      }
-    }
-
-    // add the LaTeX expression from double dollars:
-    const latexContent = match[1]
-    parts.push({ type: 'latex', content: latexContent })
-    lastIndex = match.index + match[0].length
-  }
-
-  // add remaining text after the last LaTeX expression:
-  if (lastIndex < input.length) {
-    const remainingText = input.slice(lastIndex)
-    if (remainingText) {
-      parts.push({ type: 'text', content: remainingText })
-    }
-  }
-
-  // if no LaTeX found, return the original text as a single text part:
-  if (!parts.length) {
-    parts.push({ type: 'text', content: input })
-  }
-
-  return parts
 }
